+++ src/PostgreSql/internal/Set-AzPostgreSqlServerAdministrator.ps1
@@ -39,9 +39,9 @@ COMPLEX PARAMETER PROPERTIES
 To create the parameters described below, construct a hash table containing the appropriate properties. For information on hash tables, run Get-Help about_Hash_Tables.
 
 PROPERTY <IServerAdministratorResource>: Represents a and external administrator to be created.
-  Login <String>: The server administrator login account name.
-  Sid <String>: The server administrator Sid (Secure ID).
-  TenantId <String>: The server Active Directory Administrator tenant id.
+  [Login <String>]: The server administrator login account name.
+  [Sid <String>]: The server administrator Sid (Secure ID).
+  [TenantId <String>]: The server Active Directory Administrator tenant id.
 .Link
 https://docs.microsoft.com/en-us/powershell/module/az.postgresql/set-azpostgresqlserveradministrator
 #>